feat: implement GitHub webhook event processing
Add comprehensive webhook event processing for workflow_run and workflow_job events with database persistence and error handling.

Features:
- WebhookHandler module with upsert logic for repositories, workflow runs, and jobs
- Updated database schema to use bigint for GitHub IDs
- Robust error handling with proper logging and validation
- Comprehensive test coverage (23 new tests)
- Integration with existing webhook controller

+++ priv/repo/migrations/20250803033536_create_repositories.exs
@@ -5,7 +5,7 @@ defmodule CiRunners.Repo.Migrations.CreateRepositories do
     create table(:repositories) do
       add :owner, :string, null: false
       add :name, :string, null: false
-      add :github_id, :integer, null: false
+      add :github_id, :bigint, null: false
 
       timestamps(type: :utc_datetime)
     end
